Fairly major (conceptually) overhaul
There are a large number of changes, all organized around a
solidification of the terminology and generalization for various
usecases.

`AbstractDataSource` needs to be implemented for each new datasource (eg
a CSV, an HDF5 file, etc.). A `readtrial` method should be added for
each datasource type.

Additionally, a "base" segment type, which is a subtype of the custom
datasource type should be added. This base segment type is what should
be returned by the `readtrial` method. It generally will just include
the data from the whole trial and any additional information (eg events)
which are considered part of the trial data (ie and not part of the
`Trial` type describing the trial as a whole.